Add 2nd-order Trotter, MD collision safety check, fix run_batch_jit dtype bug (#18)
Three fixes verified against a code-review pass over the codebase:
- trotter_evolve_ops gains order=2 (Strang splitting), quadratically
  more accurate than the existing order=1 for the same n_steps, at 2x
  gates/step -- useful when gate count directly limits noise budget
  (wormhole_syk_teleportation.py's Trotter noise experiments).
- run_md_trajectory now raises a clear RuntimeError if two atoms end
  up closer than a physically-motivated 0.3 A floor after a step,
  instead of silently feeding a near-collided geometry into Hartree-Fock
  (which diverges rather than failing informatively) -- catches the
  actual failure mode of a too-large dt_fs.
- run_batch_jit's init_sv was hardcoded to complex128 regardless of
  self.dtype/use_float32, silently ignoring the instance's own
  configured precision (a separate instance of the same bug class
  run_circuit_jit already had a documented fix for). Now derives from
  self.dtype like the rest of the JIT path already does.

124 tests pass (15 trotter, 18 qmmm, 91 simulator), all new behavior
covered by new tests, no regressions in existing ones.

+++ dense_evolution/simulator.py
@@ -492,7 +492,13 @@ class DenseSVSimulator:
             )
 
         template    = jnp.array(compiled_ops, dtype=jnp.float64)
-        init_sv     = jnp.zeros(self.dim, dtype=jnp.complex128).at[0].set(1.0)
+        # self.dtype, not a hardcoded jnp.complex128: _apply_gate_fast_step
+        # (compiler.py) already derives its working dtype from the input
+        # statevector specifically so use_float32=True instances run in
+        # complex64 end to end -- hardcoding complex128 here silently
+        # discarded that for every run_batch_jit/run_parametric_batch_jit
+        # call regardless of the instance's own configured dtype.
+        init_sv     = jnp.zeros(self.dim, dtype=self.dtype).at[0].set(1.0)
 
         def simulate_single_instance(single_params: "jnp.ndarray") -> "jnp.ndarray":
             """Run one parameter vector through the circuit."""
